requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
PyPDF4==1.27.0
click==7.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.1.2
PyMuPDF
fitz
